chore(telemetry): moving noopLogger to separate file
